// frontend/src/lib/proto/salesorder/v1/announcement_pb.d.ts
export declare type Announcement = Message<"salesorder.v1.Announcement"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * 空 = 全系統公告
   *
   * @generated from field: string company_id = 2;
   */
  companyId: string;

  /**
   * 空 = 全系統或公司層公告
   *
   * @generated from field: string department_id = 3;
   */
  departmentId: string;

  /**
   * banner / news / article
   *
   * @generated from field: string type = 4;
   */
  type: string;

  /**
   * @generated from field: string title = 5;
   */
  title: string;

  /**
   * @generated from field: string content = 6;
   */
  content: string;

  /**
   * 可空
   *
   * @generated from field: string image_url = 7;
   */
  imageUrl: string;

  /**
   * 可空(banner 點擊導向)
   *
   * @generated from field: string link_url = 8;
   */
  linkUrl: string;

  /**
   * RFC3339
   *
   * @generated from field: string publish_at = 9;
   */
  publishAt: string;

  /**
   * RFC3339;空 = 不自動下架
   *
   * @generated from field: string unpublish_at = 10;
   */
  unpublishAt: string;

  /**
   * @generated from field: int32 sort_order = 11;
   */
  sortOrder: number;

  /**
   * @generated from field: bool is_active = 12;
   */
  isActive: boolean;

  /**
   * 平台投放:Web 中台
   *
   * @generated from field: bool deploy_web = 13;
   */
  deployWeb: boolean;

  /**
   * 平台投放:App 首頁
   *
   * @generated from field: bool deploy_app = 14;
   */
  deployApp: boolean;

  /**
   * RFC3339
   *
   * @generated from field: string created_at = 15;
   */
  createdAt: string;

  /**
   * RFC3339
   *
   * @generated from field: string updated_at = 16;
   */
  updatedAt: string;
};